requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
idna==3.6
isodate==0.6.1
lxml==5.1.0
more-itertools==10.2.0
numpy==1.26.4
pandas==2.2.0
pendulum==3.0.0
platformdirs==4.2.0
pyarrow==15.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
simple-salesforce==1.12.5
six==1.16.0
time-machine==2.13.0
TogglPy==0.1.2
tzdata==2024.1
urllib3==2.2.1
zeep==4.2.1
